bug_fix_clarity_calc_arkane_rate.py
By setting the atom corrections to false for rate calculations, it provides better computational predictions for rates. This is because atom corrections can not predict transition states accurately. 

The negative frequency needs to be multiplied by the frequency scaling factor to correct for the systematic errors by the computational methods.

For clarity, I define E0_reac, E0_prod, and E0_ts.